requirements.txt
paho-mqtt
pysmartthings
